Rename to Flask-Babel2
