+++ README.md
@@ -1,13 +1,13 @@
 # Flask Babel - 0.12.2
 
-[![Build Status](https://travis-ci.org/python-babel/flask-babel.svg?branch=master)](https://travis-ci.org/python-babel/flask-babel)
-[![PyPI](https://img.shields.io/pypi/v/flask-babel.svg?maxAge=2592000)](https://pypi.python.org/pypi/Flask-Babel)
-![license](https://img.shields.io/github/license/python-babel/flask-babel.svg?maxAge=2592000)
-![PyPI - Python Version](https://img.shields.io/pypi/pyversions/flask-babel.svg)
+[![Build Status](https://travis-ci.org/hasgeek/flask-babel2.svg?branch=master)](https://travis-ci.org/hasgeek/flask-babel2)
+![license](https://img.shields.io/github/license/hasgeek/flask-babel2.svg?maxAge=2592000)
 
 Implements i18n and l10n support for Flask.  This is based on the Python
 [babel][] module as well as [pytz][] both of which are installed automatically
 for you if you install this library.
+
+This release merges additional functionality from Flask-BabelEx.
 
 # Documention
 
@@ -15,4 +15,4 @@ The latest documentation is available [here][docs].
 
 [babel]: https://github.com/python-babel/babel
 [pytz]: https://pypi.python.org/pypi/pytz/
-[docs]: https://pythonhosted.org/Flask-Babel/
+[docs]: https://pythonhosted.org/Flask-Babel2/
