+++ flask_babel2/_compat.py
